backend/requirements.txt
flask==2.3.3
flask-cors==4.0.0
scikit-learn==1.3.1
pandas==2.0.3
numpy==1.24.3
joblib==1.3.1
python-dotenv==1.0.0
pytest==7.4.3
pytest-cov==4.1.0
